# Conflicts:
#	Translations/WinMerge/Galician.po

+++ Src/FileFilterMgr.cpp
@@ -170,7 +170,8 @@ FileFilter * FileFilterMgr::LoadFilterFile(const String& szFilepath, int & error
 	}
 
 	file.ReadBom(); // in case it is a Unicode file, let UniMemFile handle BOM
-	if (!file.IsUnicode() && !ucr::CheckForInvalidUtf8(reinterpret_cast<const char*>(file.GetBase()), file.GetFileSize()))
+	if (!file.IsUnicode() && !ucr::CheckForInvalidUtf8(
+		reinterpret_cast<const char*>(file.GetBase()), static_cast<size_t>(file.GetFileSize())))
 		file.SetUnicoding(ucr::UTF8);
 
 	String fileName;
